update: implement Comment Modal
