+++ components/BlogCard.tsx
@@ -3,7 +3,7 @@ import { Blog } from "@/models/UserModel";
 
 import LikeButton from "./LikeButton";
 
-import React from "react";
+import React, { useEffect } from "react";
 import { Card, CardBody, CardHeader, CardFooter } from "@nextui-org/card";
 import { FaRegComment } from "react-icons/fa";
 import { useDisclosure } from "@nextui-org/react";
@@ -13,10 +13,12 @@ import CommentModal from "./CommentModal";
 interface CardProps {
   blog: Blog;
   handleLike: (blog: Blog) => void;
+  handleComment: (blog: Blog) => void;
 }
 
-const BlogCard = ({ blog, handleLike }: CardProps) => {
+const BlogCard = ({ blog, handleLike, handleComment }: CardProps) => {
   const { isOpen, onOpen, onOpenChange } = useDisclosure();
+  console.log(blog);
   return (
     <>
       <Card className="w-full p-2 m-4 bg-theme text-cyan-50 font-semibold  drop-shadow-2xl ">
@@ -42,6 +44,7 @@ const BlogCard = ({ blog, handleLike }: CardProps) => {
         blog={blog}
         isOpen={isOpen}
         onOpenChange={onOpenChange}
+        handleComment={handleComment}
       />
     </>
   );
