+++ components/BlogList.tsx
@@ -10,6 +10,7 @@ import { Blog } from "@/models/UserModel";
 import noContent from "@/assets/No data-pana.svg";
 import BlogCard from "./BlogCard";
 import { useBlogStore } from "@/utils/store";
+import toast from "react-hot-toast";
 
 interface BlogListProps {
   blogs: Blog[];
@@ -41,6 +42,38 @@ const BlogList = ({ blogs }: BlogListProps) => {
           body: JSON.stringify(form),
         });
         const data = await response.json();
+      } catch (err) {
+        console.log(err);
+        blogStore.addBlogs(blogStore.blogs);
+      }
+    }
+  };
+
+  const handleComment = async (tempBlog: Blog) => {
+    if (!session) {
+      router.push("/signin");
+    } else {
+      try {
+        let newBlogs = blogs.map((b: Blog) => {
+          if (b._id === tempBlog._id) {
+            return { ...b, comments: tempBlog.comments };
+          }
+          return b;
+        });
+        blogStore.addBlogs(newBlogs);
+        const form = {
+          session: session,
+          blog: tempBlog,
+        };
+        const response = await fetch("/api/user", {
+          method: "PUT",
+          body: JSON.stringify(form),
+        });
+        if (response.ok) {
+          toast.success("Comment posted!");
+        } else {
+          toast.error("Error posting comment!");
+        }
       } catch (err) {
         console.log(err);
         blogStore.addBlogs(blogStore.blogs);
@@ -82,6 +115,7 @@ const BlogList = ({ blogs }: BlogListProps) => {
             <BlogCard
               blog={blog}
               handleLike={handleLike}
+              handleComment={handleComment}
             />
           </li>
         ))}
